+++ src/Bootstrap/Context/Context.ts
@@ -25,7 +25,7 @@ namespace LiteMol.Bootstrap {
         dispatcher = new Service.Dispatcher();
         logger = new Service.Logger(this);
         performance = new Core.Utils.PerformanceMonitor();
-        scene: Visualization.SceneWrapper = <any>void 0; // injected by the Viewpoer component.        
+        scene: Visualization.SceneWrapper = <any>void 0; // injected by the Viewport component.        
         tree = Tree.create<Entity.Any>(this, Entity.Root.create(Entity.RootTransform, { label: 'Root Entity' }));        
         currentEntity: Entity.Any | undefined = void 0;        
         transforms = new TransformManager(this);